membuat komponen pada payment
